Fix Windows builds

+++ include/qtensornetwork.hpp
@@ -414,7 +414,7 @@ public:
                             bi_or_ip(&result, pow2(i));
                         }
                     }
-                    shotsArray[shot] = result;
+                    shotsArray[(size_t)shot] = result;
                 });
             }
 #endif
